Handle server hangup when polling for Net BIOS packet
The smb.SMBConnection.connect() method was getting stuck in an infinite 'while' loop and consuming 100% CPU.

This happens if the server hangs up the connection because the client keeps reading empty strings from the socket.
In our case, we were trying to connect to an IP address which exists and accepts connections on port 139, but is not set up as a proper SMB server.